+++ pkg/connector/client.go
@@ -53,6 +53,7 @@ func (wa *WhatsAppConnector) LoadUserLogin(ctx context.Context, login *bridgev2.
 		directMediaRetries: make(map[networkid.MessageID]*directMediaRetry),
 		mediaRetryLock:     semaphore.NewWeighted(wa.Config.HistorySync.MediaRequests.MaxAsyncHandle),
 		pushNamesSynced:    exsync.NewEvent(),
+		createDedup:        exsync.NewSet[types.MessageID](),
 	}
 	login.Client = w
 
@@ -116,6 +117,7 @@ type WhatsAppClient struct {
 	isNewLogin         bool
 	pushNamesSynced    *exsync.Event
 	lastPresence       types.Presence
+	createDedup        *exsync.Set[types.MessageID]
 }
 
 var (
@@ -421,7 +423,7 @@ func (wa *WhatsAppClient) HandleMatrixViewingChat(ctx context.Context, msg *brid
 	// Reset, but don't save, portal last sync time for immediate sync now
 	msg.Portal.Metadata.(*waid.PortalMetadata).LastSync.Time = time.Time{}
 	// Enqueue for the sync, don't block on it completing
-	wa.EnqueuePortalResync(msg.Portal)
+	wa.EnqueuePortalResync(msg.Portal, true)
 
 	if msg.Portal.OtherUserID != "" {
 		// If this is a DM, also sync the ghost of the other user immediately