+++ pkg/msgconv/from-whatsapp.go
@@ -33,7 +33,6 @@ import (
 	"go.mau.fi/whatsmeow/types"
 	_ "golang.org/x/image/webp"
 	"maunium.net/go/mautrix/bridgev2"
-	"maunium.net/go/mautrix/bridgev2/database"
 	"maunium.net/go/mautrix/bridgev2/networkid"
 	"maunium.net/go/mautrix/event"
 	"maunium.net/go/mautrix/id"
@@ -138,6 +137,7 @@ func (mc *MessageConverter) ToMatrix(
 	rawWaMsg *waE2E.Message,
 	info *types.MessageInfo,
 	isViewOnce bool,
+	isBackfill bool,
 	previouslyConvertedPart *bridgev2.ConvertedMessagePart,
 ) *bridgev2.ConvertedMessage {
 	ctx = context.WithValue(ctx, contextKeyClient, client)
@@ -213,7 +213,7 @@ func (mc *MessageConverter) ToMatrix(
 	case waMsg.GroupInviteMessage != nil:
 		part, contextInfo = mc.convertGroupInviteMessage(ctx, info, waMsg.GroupInviteMessage)
 	case waMsg.ProtocolMessage != nil && waMsg.ProtocolMessage.GetType() == waE2E.ProtocolMessage_EPHEMERAL_SETTING:
-		part, contextInfo = mc.convertEphemeralSettingMessage(ctx, waMsg.ProtocolMessage)
+		part, contextInfo = mc.convertEphemeralSettingMessage(ctx, waMsg.ProtocolMessage, info.Timestamp, isBackfill)
 	case waMsg.EncCommentMessage != nil:
 		part = failedCommentPart
 	default:
@@ -240,10 +240,16 @@ func (mc *MessageConverter) ToMatrix(
 	}
 	if contextInfo.GetExpiration() > 0 {
 		cm.Disappear.Timer = time.Duration(contextInfo.GetExpiration()) * time.Second
-		cm.Disappear.Type = database.DisappearingTypeAfterRead
+		cm.Disappear.Type = event.DisappearingTypeAfterSend
 	}
 	if portal.Disappear.Timer != cm.Disappear.Timer && portal.Metadata.(*waid.PortalMetadata).DisappearingTimerSetAt < contextInfo.GetEphemeralSettingTimestamp() {
-		portal.UpdateDisappearingSetting(ctx, cm.Disappear, intent, info.Timestamp, true, true)
+		portal.UpdateDisappearingSetting(ctx, cm.Disappear, bridgev2.UpdateDisappearingSettingOpts{
+			Sender:     intent,
+			Timestamp:  info.Timestamp,
+			Implicit:   true,
+			Save:       true,
+			SendNotice: true,
+		})
 	}
 	if contextInfo.GetStanzaID() != "" {
 		pcp, _ := types.ParseJID(contextInfo.GetParticipant())