+++ pkg/msgconv/wa-misc.go
@@ -27,6 +27,7 @@ import (
 	"github.com/rs/zerolog"
 	"go.mau.fi/util/exerrors"
 	"go.mau.fi/util/ptr"
+	"go.mau.fi/whatsmeow/proto/waAICommon"
 	"go.mau.fi/whatsmeow/proto/waE2E"
 	"go.mau.fi/whatsmeow/types"
 	"google.golang.org/protobuf/proto"
@@ -114,11 +115,11 @@ func (mc *MessageConverter) convertGroupInviteMessage(ctx context.Context, info 
 	}, msg.GetContextInfo()
 }
 
-func (mc *MessageConverter) convertEphemeralSettingMessage(ctx context.Context, msg *waE2E.ProtocolMessage) (*bridgev2.ConvertedMessagePart, *waE2E.ContextInfo) {
+func (mc *MessageConverter) convertEphemeralSettingMessage(ctx context.Context, msg *waE2E.ProtocolMessage, ts time.Time, isBackfill bool) (*bridgev2.ConvertedMessagePart, *waE2E.ContextInfo) {
 	portal := getPortal(ctx)
 	portalMeta := portal.Metadata.(*waid.PortalMetadata)
 	disappear := database.DisappearingSetting{
-		Type:  database.DisappearingTypeAfterRead,
+		Type:  event.DisappearingTypeAfterSend,
 		Timer: time.Duration(msg.GetEphemeralExpiration()) * time.Second,
 	}
 	if disappear.Timer == 0 {
@@ -126,26 +127,39 @@ func (mc *MessageConverter) convertEphemeralSettingMessage(ctx context.Context, 
 	}
 	dontBridge := portal.Disappear == disappear
 	content := bridgev2.DisappearingMessageNotice(disappear.Timer, false)
-	if msg.EphemeralSettingTimestamp == nil || portalMeta.DisappearingTimerSetAt < msg.GetEphemeralSettingTimestamp() {
-		portal.Disappear = disappear
-		portalMeta.DisappearingTimerSetAt = msg.GetEphemeralSettingTimestamp()
-		err := portal.Save(ctx)
-		if err != nil {
-			zerolog.Ctx(ctx).Err(err).Msg("Failed to save portal after updating expiration timer")
+	if !isBackfill {
+		if msg.EphemeralSettingTimestamp == nil || portalMeta.DisappearingTimerSetAt < msg.GetEphemeralSettingTimestamp() {
+			portalMeta.DisappearingTimerSetAt = msg.GetEphemeralSettingTimestamp()
+			portal.UpdateDisappearingSetting(ctx, disappear, bridgev2.UpdateDisappearingSettingOpts{
+				Sender:     getIntent(ctx),
+				Timestamp:  ts,
+				Implicit:   false,
+				Save:       true,
+				SendNotice: false,
+			})
+		} else {
+			content.Body += ", but the change was ignored."
 		}
-	} else {
-		content.Body += ", but the change was ignored."
 	}
 	return &bridgev2.ConvertedMessagePart{
-		Type:       event.EventMessage,
-		Content:    content,
+		Type:    event.EventMessage,
+		Content: content,
+		Extra: map[string]any{
+			"com.beeper.action_message": map[string]any{
+				"type":       "disappearing_timer",
+				"timer":      disappear.Timer.Milliseconds(),
+				"timer_type": disappear.Type,
+				"implicit":   false,
+				"backfill":   isBackfill,
+			},
+		},
 		DontBridge: dontBridge,
 	}, nil
 }
 
 const eventMessageTemplate = `
 {{- if .Name -}}
-	<h4>{{ .Name }}</h4>
+	<h4>{{ .Name }} {{- if .IsCanceled -}}<span> (Canceled)</span>{{- end -}}</h4>
 {{- end -}}
 {{- if .StartTime -}}
 	<p>
@@ -171,6 +185,7 @@ var eventMessageTplParsed = exerrors.Must(template.New("eventmessage").Parse(str
 
 type eventMessageParams struct {
 	Name            string
+	IsCanceled      bool
 	JoinLink        string
 	StartTimeISO    string
 	StartTime       string
@@ -183,6 +198,7 @@ type eventMessageParams struct {
 func (mc *MessageConverter) convertEventMessage(ctx context.Context, msg *waE2E.EventMessage) (*bridgev2.ConvertedMessagePart, *waE2E.ContextInfo) {
 	params := &eventMessageParams{
 		Name:            msg.GetName(),
+		IsCanceled:      msg.GetIsCanceled(),
 		JoinLink:        msg.GetJoinLink(),
 		Location:        msg.GetLocation().GetName(),
 		DescriptionHTML: template.HTML(parseWAFormattingToHTML(msg.GetDescription(), false)),
@@ -249,7 +265,7 @@ func (mc *MessageConverter) convertRichResponseMessage(ctx context.Context, msg 
 	var body strings.Builder
 
 	for i, submsg := range msg.GetSubmessages() {
-		if submsg.GetMessageType() == waE2E.AIRichResponseMessage_AI_RICH_RESPONSE_TEXT {
+		if submsg.GetMessageType() == waAICommon.AIRichResponseSubMessageType_AI_RICH_RESPONSE_TEXT {
 			if i > 0 {
 				body.WriteString("\n")
 			}